Keep the VM state just before an error in the debugger
Because we are juggling VM states in vm_history for the time travel
debugger, naively returning an error from the step execution will remove
the latest VM from the history, resulting a state at the step one before
the user would expect. We keep the error in a result variable and raise
error at appropriate time to keep the latest VM in the history.