Return Result from TypeSet::try_intersect to immitate previous behavior
If the type set intersection become an empty set, we find a
type check error, but it won't communicate what was wrong about the
code. We return error early if we detects an invalid intersection of
type sets, for example, an array with incompatible size, because these
kinds of type errors are hard to find out without good compile error
messages.

Previous implementation of type checker was actually pretty good at it,
so we want to retain that property. Thankfully, we have a bunch of unit
tests with expected error messages, so we can adapt type inferer's
behavior to give similar messages.