Keep track of variable reference for type inference
lvalues can be an array element or a tuple element, in which cases we need
to keep track of the reference to the element type, which we will use to
modify by the inferred type. This is necessary for assignment
expressions such as `a[0] = 1` or `a.0 = 1` to participate in the type
inference.