BREAKING CHANGE: Semicolon affects type inference now
We ignored the presence of the semicolon after the last statement in a
function or a brace expression, but now it matters for type
compatibility.

We need this change because we need definite types for functions.
For now, we do not infer return types, but wheter we infer or not,
we need a definite type. The return type should match the last
expression statement if there is no semicolon, or Void if there is.
So it matters whether the semicolon exists in AST.

We also mandated return type declaration of functions. If you do not
declare a return type, it means void, not any.

It also affects how the bytecode is generated, so type
inference/checking is no longer optional.